add working test_ml_pipeline.py with logs for 1epoch in CPU (5.3mins for train), (1.2mins for eval) #8
